# swagger/swagger.yaml
consumes:
    - application/json
definitions:
...
    CreateDingTalkConfigResponse:
        description: CreateDingTalkConfigResponse 创建钉钉配置列表的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    CreateDingtalkConfigRequest:
        description: CreateDingtalkConfigRequest 创建钉钉配置列表的请求数据结构
        properties:
            Name:
                description: 机器人名称
                type: string
            Url:
                description: 机器人webhoook地址
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    DeleteDingTalkConfigResponse:
        description: DeleteDingTalkConfigResponse 删除钉钉配置列表的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    DeleteDingtalkConfigRequest:
        description: DeleteDingtalkConfigRequest 删除钉钉配置列表的请求数据结构
        properties:
            Id:
                description: 配置id
                format: int64
                type: integer
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    DingtalkConfig:
        description: DingtalkConfig 钉钉配置
        properties:
            Id:
                format: int64
                type: integer
            Name:
                description: 机器人名称
                type: string
            Url:
                description: 机器人webhoook地址
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
    DingtalkConfigList:
        properties:
            Items:
                items:
                    $ref: '#/definitions/DingtalkConfig'
                type: array
            Total:
                format: int64
                type: integer
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/model
    ErrResponse:
        description: ErrResponse 错误响应
        properties:
            Code:
                example: 1001
                format: int64
                type: integer
            Data:
                example: {}
                type: object
            Message:
                example: 失败
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    GetFeishuConfigResponse:
        description: GetFeishuConfigResponse 获取飞书配置的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                $ref: '#/definitions/GetFeishuConfigResponseData'
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    GetFeishuConfigResponseData:
        properties:
            Url:
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    GetMailboxConfigResponse:
        description: GetMailboxConfigResponse 获取邮箱配置的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                $ref: '#/definitions/GetMailboxConfigResponseData'
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    GetMailboxConfigResponseData:
        properties:
            DisplayName:
                type: string
            Port:
                format: int64
                type: integer
            SSL:
                type: boolean
            SmtpServer:
                type: string
                x-go-name: SMTPServer
            Template:
                description: Password    string `json:"Password"` // 不返回密码
                type: string
            Username:
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    GetSMSConfigResponse:
        description: GetSMSConfigResponse 获取短信配置的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                $ref: '#/definitions/GetSMSConfigResponseData'
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    GetSMSConfigResponseData:
        properties:
            AK:
                type: string
            Ip:
                type: string
                x-go-name: IP
            Protocol:
                type: string
            SK:
                type: string
            Sign:
                type: string
            Template:
                type: string
            Url:
                type: string
                x-go-name: URL
            User:
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    GetWechatConfigResponse:
        description: GetWechatConfigResponse 获取企业微信配置的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                $ref: '#/definitions/GetWechatConfigResponseData'
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    GetWechatConfigResponseData:
        properties:
            AppId:
                type: string
            AppSecret:
                type: string
            CompanyId:
                type: string
            Template:
                type: string
            ToParty:
                items:
                    type: string
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    ListDingTalkConfigResponse:
        description: ListDingTalkConfigResponse 获取钉钉配置列表的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                $ref: '#/definitions/DingtalkConfigList'
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    ListDingtalkConfigRequest:
        description: ListDingtalkConfigRequest 获取钉钉配置列表的请求数据结构
        properties:
            PageNo:
                description: 当前页码，从1开始
                format: int64
                type: integer
            PageSize:
                description: 每页数据个数，默认10个, 当值为-1时不分页
                format: int64
                type: integer
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    UpdateDingTalkConfigResponse:
        description: UpdateDingTalkConfigResponse 创建钉钉配置列表的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdateDingtalkConfigRequest:
        description: UpdateDingtalkConfigRequest 创建钉钉配置列表的请求数据结构
        properties:
            Id:
                format: int64
                type: integer
            Name:
                description: 机器人名称
                type: string
            Url:
                description: 机器人webhoook地址
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdateFeishuConfigRequest:
        description: UpdateFeishuConfigRequest 更新飞书远程配置的请求数据结构
        properties:
            Url:
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdateFeishuConfigResponse:
        description: UpdateFeishuConfigResponse 更新飞书远程配置的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdateMailboxConfigRequest:
        description: UpdateMailboxConfigRequest 更新邮箱远程配置的请求数据结构
        properties:
            DisplayName:
                type: string
            Password:
                type: string
            Port:
                format: int64
                type: integer
            SSL:
                type: boolean
            SmtpServer:
                type: string
                x-go-name: SMTPServer
            Template:
                type: string
            Username:
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdateMailboxConfigResponse:
        description: UpdateMailboxConfigResponse 更新邮箱远程配置的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    UpdateSMSConfigRequest:
        description: UpdateSMSConfigRequest 更新短信远程配置的请求数据结构
        properties:
            AK:
                type: string
            Ip:
                type: string
                x-go-name: IP
            Protocol:
                type: string
            SK:
                type: string
            Sign:
                type: string
            Template:
                type: string
            Url:
                type: string
                x-go-name: URL
            User:
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdateSMSConfigResponse:
        description: UpdateSMSConfigResponse 更新短信远程配置的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdateWechatConfigRequest:
        description: UpdateWechatConfigRequest 更新企业微信远程配置的请求数据结构
        properties:
            AppId:
                type: string
            AppSecret:
                type: string
            CompanyId:
                type: string
            Template:
                type: string
            ToParty:
                items:
                    type: string
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
    UpdateWechatConfigResponse:
        description: UpdateWechatConfigResponse 更新企业微信远程配置的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
host: xxxxx.swagger.galaxy.cloud
info:
    description: 告警策略
    title: 告警策略
    version: 1.0.0
paths:
...
    /alarm/v2/remoteconfig/dingtalk/create:
        post:
            description: 创建钉钉配置
            operationId: createDingtalkConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/CreateDingtalkConfigRequest'
            responses:
                "200":
                    $ref: '#/responses/createDingtalkConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 创建钉钉配置
            tags:
                - remote_config
    /alarm/v2/remoteconfig/dingtalk/delete:
        post:
            description: 删除钉钉配置
            operationId: deleteDingtalkConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/DeleteDingtalkConfigRequest'
            responses:
                "200":
                    $ref: '#/responses/deleteDingtalkConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 删除钉钉配置
            tags:
                - remote_config
    /alarm/v2/remoteconfig/dingtalk/list:
        post:
            description: 获取钉钉配置列表
            operationId: listDingtalkConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/ListDingtalkConfigRequest'
            responses:
                "200":
                    $ref: '#/responses/listDingtalkConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 获取钉钉配置列表
            tags:
                - remote_config
    /alarm/v2/remoteconfig/dingtalk/update:
        post:
            description: 更新钉钉配置
            operationId: updateDingtalkConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdateDingtalkConfigRequest'
            responses:
                "200":
                    $ref: '#/responses/updateDingtalkConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 更新钉钉配置
            tags:
                - remote_config
    /alarm/v2/remoteconfig/feishu:
        get:
            description: 获取飞书配置
            operationId: getFeishuConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
            responses:
                "200":
                    $ref: '#/responses/getFeishuConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 获取飞书配置
            tags:
                - remote_config
        post:
            description: 更新飞书配置
            operationId: updateFeishuConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdateFeishuConfigRequest'
            responses:
                "200":
                    $ref: '#/responses/updateFeishuConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 更新飞书配置
            tags:
                - remote_config
    /alarm/v2/remoteconfig/mailbox:
        get:
            description: 获取邮箱配置
            operationId: getMailboxConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
            responses:
                "200":
                    $ref: '#/responses/getMailboxConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 获取邮箱配置
            tags:
                - remote_config
        post:
            description: 更新邮箱配置
            operationId: updateMailboxConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdateMailboxConfigRequest'
            responses:
                "200":
                    $ref: '#/responses/updateMailboxConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 更新邮箱配置
            tags:
                - remote_config
    /alarm/v2/remoteconfig/sms:
        get:
            description: 获取短信配置
            operationId: getSMSConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
            responses:
                "200":
                    $ref: '#/responses/getSMSConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 获取短信配置
            tags:
                - remote_config
        post:
            description: 更新短信配置
            operationId: updateSMSConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdateSMSConfigRequest'
            responses:
                "200":
                    $ref: '#/responses/updateSMSConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 更新短信配置
            tags:
                - remote_config
    /alarm/v2/remoteconfig/wechat:
        get:
            description: 获取企业微信配置
            operationId: getWechatConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
            responses:
                "200":
                    $ref: '#/responses/getWechatConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 获取企业微信配置
            tags:
                - remote_config
        post:
            description: 更新企业微信配置
            operationId: updateWechatConfigRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdateWechatConfigRequest'
            responses:
                "200":
                    $ref: '#/responses/updateWechatConfigResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 更新企业微信配置
            tags:
                - remote_config
...
produces:
    - application/json
responses:
...
    createDingtalkConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/CreateDingTalkConfigResponse'
...
    deleteDingtalkConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/DeleteDingTalkConfigResponse'
...
    errResponse:
        description: 发生错误时的返回信息
        schema:
            $ref: '#/definitions/ErrResponse'
    getFeishuConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/GetFeishuConfigResponse'
    getMailboxConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/GetMailboxConfigResponse'
...
    getSMSConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/GetSMSConfigResponse'
    getWechatConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/GetWechatConfigResponse'
...
    listDingtalkConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/ListDingTalkConfigResponse'
...
    updateDingtalkConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/UpdateDingTalkConfigResponse'
    updateFeishuConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/UpdateFeishuConfigResponse'
    updateMailboxConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/UpdateMailboxConfigResponse'
...
    updateSMSConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/UpdateSMSConfigResponse'
    updateWechatConfigResponse:
        description: ""
        schema:
            $ref: '#/definitions/UpdateWechatConfigResponse'
schemes:
    - http
security:
    - api_key: []
securityDefinitions:
    api_key:
        in: header
        name: api_key
        type: apiKey
    basic:
        type: basic
swagger: "2.0"